webhooks/views.py
import json
from rest_framework import views, response, status
from django.conf import settings
from django.core.mail import send_mail
from django.template.loader import render_to_string
from webhooks.models import Webhook
from webhooks.messages import outflow_message
from services.callmebot import CallMeBot


class WebhookOrderView(views.APIView):

    def post(self, request):
        data = request.data

        Webhook.objects.create(
            event_type=data.get('event_type'),
            event=json.dumps(data, ensure_ascii=False),
        )

        product_name = data.get('product')
        quantity = data.get('quantity')
        product_cost_price = data.get('product_cost_price')
        product_sale_price = data.get('product_sale_price')
        product_description = data.get('description')
        total_value = product_sale_price * quantity
        profit_value = total_value - (product_cost_price * quantity)
        
        message = outflow_message.format(
            product_name, 
            quantity, 
            total_value,
            profit_value,
            product_description
        )
        callmebot = CallMeBot()
        callmebot.send_message(message)

        # data['total_value'] = total_value
        # data['profit_value'] = profit_value
        # send_mail(
        #     subject='Nova Saída (SGE)',
        #     message='',
        #     from_email=f'SGE <{settings.EMAIL_HOST_USER}>',
        #     recipient_list=[settings.EMAIL_ADMIN_RECEIVER],
        #     html_message=render_to_string('outflow.html', data),
        #     fail_silently=False,
        # )

        return response.Response(
            data=data,
            status=status.HTTP_200_OK,
        )
